% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salaries.R
\name{get_salaries}
\alias{get_salaries}
\title{get_salaries}
\usage{
get_salaries(groups = "all")
}
\arguments{
\item{groups}{character vector with the groups listed here: https://www.tbs-sct.canada.ca/pubs_pol/hrpubs/coll_agre/rates-taux-eng.asp Note that this function only supports select groups. Use "all" (default) to download all supported groups}
}
\value{
dataframe with public service pay rates
}
\description{
This function gets the publicly available DFO salary information from
https://www.tbs-sct.canada.ca/pubs_pol/hrpubs/coll_agre/rates-taux-eng.asp.
}
\details{
Disclaimer: If you're obtaining salary information for group="TC", subgroup=TI
this function obtains the first classification tables and ignores ones that are
specific to Aviation, Marine, and Railway Safety.
}
\examples{
# getting salaries for the SP (Applied Science and Patent Examination) group
SP_salaries <- get_salaries("SP")

\dontrun{
# getting all supported salaries
all_salaries <- get_salaries()}
}
