% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{TBS groups and classifications}
\format{
A data frame with 284 rows and 2 variables:
\describe{
\item{\code{Group}}{character TBS groups that can be used in \code{get_salaries()}}
\item{\code{Classification}}{character The classifications found in each group}
}
}
\usage{
groups
}
\description{
The groups and classifications that this package can currently extract from the TBS website.
}
\details{
R script to download/create the data is in \href{https://github.com/dfo-mar-odis/TBSpayRates/blob/main/data-raw/groups.R}{data-raw/groups.R}
}
\keyword{datasets}
