% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBSpayRates-package.R
\docType{package}
\name{TBSpayRates-package}
\alias{TBSpayRates}
\alias{TBSpayRates-package}
\title{TBSpayRates: Rates of pay for the public service from the Treasury Board of Canada Secretariat}
\description{
A package to extract rates of pay for public servants from the Treasury Board of Canada Secretariat and tools for project planning/salary forecasting.
}
\keyword{internal}
